+++ src/Command/PHPUnitCommand.php
@@ -31,6 +31,9 @@ class PHPUnitCommand extends AbstractMoodleCommand
             ->setDescription('Run PHPUnit on a plugin')
             ->addOption('coverage-text', null, InputOption::VALUE_NONE, 'Generate and print code coverage report in text format')
             ->addOption('coverage-clover', null, InputOption::VALUE_NONE, 'Generate code coverage report in Clover XML format')
+            ->addOption('coverage-pcov', null, InputOption::VALUE_NONE, 'Use the pcov extension to calculate code coverage')
+            ->addOption('coverage-xdebug', null, InputOption::VALUE_NONE, 'Use the xdebug extension to calculate code coverage')
+            ->addOption('coverage-phpdbg', null, InputOption::VALUE_NONE, 'Use the phpdbg binary to calculate code coverage')
             ->addOption('fail-on-incomplete', null, InputOption::VALUE_NONE, 'Treat incomplete tests as failures')
             ->addOption('fail-on-risky', null, InputOption::VALUE_NONE, 'Treat risky tests as failures')
             ->addOption('fail-on-skipped', null, InputOption::VALUE_NONE, 'Treat skipped tests as failures')
@@ -52,7 +55,7 @@ class PHPUnitCommand extends AbstractMoodleCommand
         }
 
         $colors  = $output->isDecorated() ? '--colors="always"' : '';
-        $binary  = $this->resolveBinary($input);
+        $binary  = $this->resolveBinary($input, $output);
         $options = $this->resolveOptions($input);
         $process = $this->execute->passThrough(
             sprintf('%s%s/vendor/bin/phpunit %s %s', $binary, $this->moodle->directory, $colors, $options),
@@ -78,6 +81,9 @@ class PHPUnitCommand extends AbstractMoodleCommand
         if ($this->supportsCoverage() && $input->getOption('coverage-clover')) {
             $options[] = sprintf('--coverage-clover %s/coverage.xml', getcwd());
         }
+        if ($input->getOption('verbose')) {
+            $options[] = '--verbose';
+        }
         foreach (['fail-on-warning', 'fail-on-risky', 'fail-on-skipped', 'fail-on-warning'] as $option) {
             if ($input->getOption($option)) {
                 $options[] = '--'.$option;
@@ -95,11 +101,12 @@ class PHPUnitCommand extends AbstractMoodleCommand
     /**
      * Use phpdbg if we are generating code coverage.
      *
-     * @param InputInterface $input
+     * @param InputInterface  $input
+     * @param OutputInterface $output
      *
      * @return string
      */
-    private function resolveBinary(InputInterface $input)
+    private function resolveBinary(InputInterface $input, OutputInterface $output)
     {
         if (!$this->supportsCoverage()) {
             return '';
@@ -108,7 +115,19 @@ class PHPUnitCommand extends AbstractMoodleCommand
             return '';
         }
 
-        return 'phpdbg -d memory_limit=-1 -qrr ';
+        // Depending on the coverage driver, selected return different values.
+        switch ($this->resolveCoverageDriver($input, $output)) {
+            case 'pcov':
+                return 'php -dxdebug.mode=off -dpcov.enabled=1 -dpcov.directory=. '; // Enable pcov, disable xdebug, just in case.
+            case 'xdebug':
+                return 'php -dpcov.enabled=0 -dxdebug.mode=coverage '; // Enable xdebug, disable pcov, just in case.
+            case 'phpdbg':
+                return 'phpdbg -d memory_limit=-1 -qrr ';
+        }
+        // No suitable coverage driver found, disabling all candidates.
+        $output->writeln('<error>No suitable driver found, disabling code coverage.</error>');
+
+        return 'php -dpcov.enabled=0 -dxdebug.mode=off ';
     }
 
     /**
@@ -119,5 +138,58 @@ class PHPUnitCommand extends AbstractMoodleCommand
     private function supportsCoverage()
     {
         return version_compare(PHP_VERSION, '7.0.0', '>=');
+    }
+
+    /**
+     * Given the current environment and options return the code coverage driver to use.
+     *
+     * @param InputInterface  $input
+     * @param OutputInterface $output
+     *
+     * @return string one of pcov, xdebug, phpdbg
+     */
+    private function resolveCoverageDriver(InputInterface $input, OutputInterface $output)
+    {
+        // Let's see if any of the coverage drivers has been forced via command line options.
+        if ($input->getOption('coverage-pcov')) {
+            // Before accepting it, perform some checks and report.
+            if (!extension_loaded('pcov')) {
+                $output->writeln('<error>PHP pcov extension not available.</error>');
+
+                return '';
+            }
+            if ($this->moodle->getBranch() < 310) {
+                $output->writeln('<error>PHP pcov coverage only can be used with Moodle 3.10 and up.</error>');
+
+                return '';
+            }
+
+            return 'pcov';
+        } elseif ($input->getOption('coverage-xdebug')) {
+            // Before accepting it, perform some checks and report.
+            if (!extension_loaded('xdebug')) {
+                $output->writeln('<error>PHP xdebug extension not available.</error>');
+
+                return '';
+            }
+
+            return 'xdebug';
+        } elseif ($input->getOption('coverage-phpdbg')) {
+            return 'phpdbg';
+        }
+
+        // Arrived here, let's find the best (pcov => xdebug => phpdbg) available driver.
+
+        if (extension_loaded('pcov') && $this->moodle->getBranch() >= 310) {
+            // If pcov is available and we are using Moodle 3.10 (PHPUnit 8.5) and up, let's use it.
+            return 'pcov';
+        }
+
+        if (extension_loaded('xdebug')) {
+            // If xdebug is available, let's use it.
+            return 'xdebug';
+        }
+
+        return 'phpdbg'; // Fallback to phpdbg (bundled with php 7.0 and up) if none of the above are available.
     }
 }